Have OmniGraffle guess if no format given on command line
Before, if you didn't pass an export format into the graffle.sh
script, it would use “pdf” as a default.  Unfortunately, it seems like
several of the format names that worked with OmniGraffle 4 no longer
work with OmniGraffle 5 — and “pdf” is one of those.  “PDF” and “PDF
vector image” (the string that appears in the Export dialog) don't
work either.

Luckily, we can leave off the “as format” portion of the “save”
AppleScript command; in this case, it seems like OmniGraffle guesses
the output format based on the extension of the output filename.  This
gives us a workaround: instead of guessing what variation of “PDF” to
use as the format name, just make sure to pass in an output filename
that ends in “.pdf”.

Closes #1